Enforce one missed property back to single-floating type

+++ osu.Game/Online/API/ArchiveDownloadRequest.cs
@@ -10,7 +10,7 @@ namespace osu.Game.Online.API
     {
         public readonly TModel Model;
 
-        public double Progress { get; private set; }
+        public float Progress { get; private set; }
 
         public event Action<float> DownloadProgressed;
